backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
openai==1.54.0
qdrant-client==1.12.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12
httpx==0.27.2
bcrypt==4.2.0
pyjwt==2.9.0
langchain==0.3.7
langchain-openai==0.2.8
langchain-qdrant==0.2.0
sentence-transformers==3.3.0
